+++ components/moduleComponents/signUp/index.tsx
@@ -2,7 +2,6 @@ import React, { use, useEffect } from "react";
 import { useForm } from "react-hook-form";
 import { FieldError } from "react-hook-form";
 import { useState } from "react";
-import axios from "@/libs/axios";
 import Image from "next/image";
 import LogoText from "@/components/atomicComponents/logoText";
 import LoginInput from "@/components/atomicComponents/loginInput";
@@ -11,7 +10,8 @@ import SnsLogin from "@/components/atomicComponents/snsLogin";
 import { email_reg, password_reg } from "@/src/utils/regexPatterns";
 import { eyeoff_svg, eyeon_svg } from "@/public/image/index";
 import styles from "./signup.module.css";
-import { useRouter } from "next/router";
+import useSignUp from "@/hook/useSignUp";
+import useEmailDuplicate from "@/hook/useEmailDuplicate";
 
 const SingUpModule = () => {
   const {
@@ -21,11 +21,12 @@ const SingUpModule = () => {
     setError,
     formState: { errors },
   } = useForm({ mode: "onBlur" });
+  
+  const getSignUp = useSignUp();
+  const isEmailDuplicate = useEmailDuplicate(setError);
 
   const [isPasswordVisible, setisPasswordVisible] = useState(false);
-  const [isPasswordConfirmVisible, setisPasswordConfirmVisible] =
-    useState(false);
-  const router = useRouter();
+  const [isPasswordConfirmVisible, setisPasswordConfirmVisible] = useState(false);
 
   const toggleisPasswordVisible = () => {
     setisPasswordVisible(!isPasswordVisible);
@@ -34,46 +35,6 @@ const SingUpModule = () => {
   const toggleisPasswordConfirmVisible = () => {
     setisPasswordConfirmVisible(!isPasswordConfirmVisible);
   };
-
-  const isEmailDuplicate = async (email: string) => {
-    try {
-      const res = await axios.post("check-email", {
-        email: email,
-      });
-    } catch (e: any) {
-      if (e.response.status === 409) {
-        setError("email", {
-          type: "manual",
-          message: "이미 사용 중인 이메일입니다.",
-        });
-      }
-    }
-  };
-
-  const getSignUp = async (data: any) => {
-    try {
-      const res = await axios.post("/sign-up", data);
-      localStorage.setItem("accessToken", res.data.accessToken);
-      router.push("/folder");
-    } catch (e: any) {
-      if (e.response.status === 400) {
-        setError("email", {
-          type: "manual",
-          message: "이메일을 확인해주세요",
-        });
-        setError("password", {
-          type: "manual",
-          message: "비밀번호를 확인해주세요",
-        });
-      }
-    }
-  };
-
-  useEffect(() => {
-    if(localStorage.getItem("accessToken")){ 
-      router.push("/folder");
-    }
-  },[])
 
   return (
     <div>
@@ -169,19 +130,3 @@ const SingUpModule = () => {
 
 export default SingUpModule;
 
-//watch는 입력된 필드의 값을 가져오는 메서드,
-//watch("password") -> password의 값을 가져옴
-//value === watch("password") ? true : "비밀번호가 일치하지 않습니다.",
-//-> 비밀번호가 같으면 true, 다르면 "비밀번호가 일치하지 않습니다."를 반환
-
-//똑같은 느김의 state 2개로 하기 싫d으니까  id에 키값을 줘서 하나로 처리하기?
-
-//오류를 수동으로 정할 때 setError를 사용한다.
-
-// React.useEffect(() => {
-//   setError("username", {
-//     type: "manual",
-//     message: "Dont Forget Your Username Should Be Cool!",
-//   })
-// }, [setError])
-//보통은 에러를 자동으로 처리하지만, 수동으로 처리할 때는 setError를 사용한다.
